move more to IO, handle a few issues with code
