+++ src/Sep/SepReader.IO.Async.cs
@@ -1,7 +1,11 @@
 ﻿//#define SYNC
+using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 #if !SYNC
+using System.Threading;
 using System.Threading.Tasks;
 #endif
 
@@ -12,14 +16,14 @@ public sealed partial class SepReader
 #if SYNC
     internal void Initialize(in SepReaderOptions options)
 #else
-    internal async ValueTask InitializeAsync(SepReaderOptions options)
+    internal async ValueTask InitializeAsync(SepReaderOptions options, CancellationToken cancellationToken)
 #endif
     {
         // Parse first row/header
 #if SYNC
         if (MoveNext())
 #else
-        if (await MoveNextAsync())
+        if (await MoveNextAsync(cancellationToken))
 #endif
         {
             A.Assert(_parsedRowsCount > 0);
@@ -96,7 +100,7 @@ public sealed partial class SepReader
 #if SYNC
     public bool MoveNext()
 #else
-    public async ValueTask<bool> MoveNextAsync()
+    public async ValueTask<bool> MoveNextAsync(CancellationToken cancellationToken)
 #endif
     {
         do
@@ -108,8 +112,139 @@ public sealed partial class SepReader
 #if SYNC
         } while (ParseNewRows());
 #else
-        } while (await ParseNewRowsAsync());
+        } while (await ParseNewRowsAsync(cancellationToken));
 #endif
         return false;
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveOptimization)]
+#if SYNC
+    internal bool ParseNewRows()
+#else
+    internal async ValueTask<bool> ParseNewRowsAsync(CancellationToken cancellationToken)
+#endif
+    {
+        _parsedRowsCount = 0;
+        _parsedRowIndex = 0;
+        _currentRowColCount = -1;
+        _currentRowColEndsOrInfosOffset = 0;
+
+#if SYNC
+        CheckPoint($"{nameof(ParseNewRows)} BEGINNING");
+#else
+        CheckPoint($"{nameof(ParseNewRowsAsync)} BEGINNING");
+#endif
+
+        // Move data to start
+        if (_parsingRowCharsStartIndex > 0)
+        {
+            A.Assert(_parser != null);
+            var offset = _chars.MoveDataToStart(ref _parsingRowCharsStartIndex, ref _charsDataEnd, _parser.PaddingLength);
+
+            // Adjust parse start
+            A.Assert(_charsParseStart >= offset);
+            _charsParseStart -= offset;
+
+            A.Assert((_parsingRowColEndsOrInfosStartIndex + _parsingRowColCount + 1) * GetIntegersPerColInfo() <= _colEndsOrColInfos.Length);
+
+            // Adjust found current row col infos, note includes col count since +1
+            if (_colUnquoteUnescape == 0)
+            {
+                ref var colEndsRef = ref GetColsRefAs<int>();
+                for (var i = 0; i <= _parsingRowColCount; i++)
+                {
+                    ref var colEnd = ref Unsafe.Add(ref colEndsRef, i + _parsingRowColEndsOrInfosStartIndex);
+                    colEnd -= offset;
+                }
+            }
+            else
+            {
+                ref var colInfosRef = ref GetColsRefAs<SepColInfo>();
+                for (var i = 0; i <= _parsingRowColCount; i++)
+                {
+                    ref var colInfo = ref Unsafe.Add(ref colInfosRef, i + _parsingRowColEndsOrInfosStartIndex);
+                    colInfo.ColEnd -= offset;
+                }
+            }
+        }
+        if (_parsingRowColEndsOrInfosStartIndex > 0)
+        {
+            var intsPerColInfo = GetIntegersPerColInfo();
+            var colInfosSpan = _colEndsOrColInfos.AsSpan();
+            var length = (_parsingRowColCount + 1) * intsPerColInfo;
+            var source = colInfosSpan.Slice(_parsingRowColEndsOrInfosStartIndex * intsPerColInfo, length);
+            var destination = colInfosSpan.Slice(0, length);
+            source.CopyTo(destination);
+            _parsingRowColEndsOrInfosStartIndex = 0;
+        }
+
+        // Ensure start conditions
+        _colEndsOrColInfos[0] = -1;
+        if (_colUnquoteUnescape != 0)
+        {
+            // QuoteCount initialize hack
+            _colEndsOrColInfos[1] = 0;
+        }
+
+        _charsDataStart = 0;
+
+        // Reset
+#if DEBUG
+        _colEndsOrColInfos.AsSpan().Slice((_parsingRowColEndsOrInfosStartIndex + _parsingRowColCount) * GetIntegersPerColInfo() + GetIntegersPerColInfo()).Fill(-42);
+#endif
+
+        var endOfFile = false;
+    LOOP:
+        CheckPoint($"{nameof(_parser)} BEFORE");
+
+        if (_parser is not null)
+        {
+            if (_colUnquoteUnescape == 0) { _parser.ParseColEnds(this); }
+            else { _parser.ParseColInfos(this); }
+        }
+
+        CheckPoint($"{nameof(_parser)} AFTER");
+        if (endOfFile)
+        {
+            CheckPoint($"{nameof(_parser)} AFTER - ENDOFFILE");
+            goto RETURN;
+        }
+
+        CheckPoint($"{nameof(_parser)} AFTER");
+
+        if (_parsedRowsCount > 0) { return true; }
+
+#if SYNC
+        endOfFile = EnsureInitializeAndReadData(endOfFile);
+#else
+        endOfFile = await EnsureInitializeAndReadDataAsync();
+#endif
+        if (endOfFile && _parsingRowCharsStartIndex < _charsDataEnd && _charsParseStart == _charsDataEnd)
+        {
+            ++_parsingRowColCount;
+            var colInfoIndex = _parsingRowColEndsOrInfosStartIndex + _parsingRowColCount;
+            if (_colUnquoteUnescape == 0)
+            {
+                _colEndsOrColInfos[colInfoIndex] = _charsDataEnd;
+            }
+            else
+            {
+                Debug.Assert(_parser is not null);
+                var quoteCount = _parser.QuoteCount;
+                Unsafe.Add(ref Unsafe.As<int, SepColInfo>(ref MemoryMarshal.GetArrayDataReference(_colEndsOrColInfos)), colInfoIndex) =
+                    new(_charsDataEnd, quoteCount);
+            }
+            ++_parsingLineNumber;
+            _parsedRows[_parsedRowsCount] = new(_parsingLineNumber, _parsingRowColCount);
+            ++_parsedRowsCount;
+
+            _parsingRowColCount = 0;
+            _parsingRowColEndsOrInfosStartIndex = colInfoIndex + 1;
+            _parsingRowCharsStartIndex = _charsDataEnd;
+            goto RETURN;
+        }
+        if (_parsedRowsCount <= 0) { goto LOOP; }
+    RETURN:
+        return _parsedRowsCount > 0;
     }
 }
