+++ src/Sep/SepReaderExtensions.Enumeration.cs
@@ -4,7 +4,9 @@ using System.Buffers;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+#if NET9_0_OR_GREATER
 using System.Threading.Tasks;
+#endif
 
 namespace nietras.SeparatedValues;
 
@@ -36,6 +38,32 @@ public static partial class SepReaderExtensions
             }
         }
     }
+
+#if NET9_0_OR_GREATER
+    public static async IAsyncEnumerable<T> EnumerateAsync<T>(this SepReader reader, Func<SepReader.Row, ValueTask<T>> select)
+    {
+        ArgumentNullException.ThrowIfNull(reader);
+        ArgumentNullException.ThrowIfNull(select);
+        await foreach (var row in reader)
+        {
+            yield return await select(row).ConfigureAwait(false);
+        }
+    }
+
+    public static async IAsyncEnumerable<T> EnumerateAsync<T>(this SepReader reader, Func<SepReader.Row, ValueTask<(bool, T)>> trySelect)
+    {
+        ArgumentNullException.ThrowIfNull(reader);
+        ArgumentNullException.ThrowIfNull(trySelect);
+        await foreach (var row in reader)
+        {
+            var (success, value) = await trySelect(row).ConfigureAwait(false);
+            if (success)
+            {
+                yield return value;
+            }
+        }
+    }
+#endif
 
     public static IEnumerable<T> ParallelEnumerate<T>(this SepReader reader, SepReader.RowFunc<T> select)
     {
@@ -186,30 +214,6 @@ public static partial class SepReaderExtensions
         foreach (var state in statesStack)
         {
             state.Dispose();
-        }
-    }
-
-    public static async IAsyncEnumerable<T> EnumerateAsync<T>(this SepReader reader, Func<SepReader.Row, ValueTask<T>> select)
-    {
-        ArgumentNullException.ThrowIfNull(reader);
-        ArgumentNullException.ThrowIfNull(select);
-        await foreach (var row in reader)
-        {
-            yield return await select(row).ConfigureAwait(false);
-        }
-    }
-
-    public static async IAsyncEnumerable<T> EnumerateAsync<T>(this SepReader reader, Func<SepReader.Row, ValueTask<(bool, T)>> trySelect)
-    {
-        ArgumentNullException.ThrowIfNull(reader);
-        ArgumentNullException.ThrowIfNull(trySelect);
-        await foreach (var row in reader)
-        {
-            var (success, value) = await trySelect(row).ConfigureAwait(false);
-            if (success)
-            {
-                yield return value;
-            }
         }
     }
 }
